Consolidate CF tests into ConditionalFormattingTest (#93)
Phase 2 added ConditionalFormatsTest as transient smoke tests for the
new API. With Phase 3 migrating ConditionalFormattingTest to the new
API, the two files overlapped in expressionRule and singleCellIsRule.

Merged the 8 unique tests from ConditionalFormatsTest into
ConditionalFormattingTest:
  - multipleRulesOnSameColumn
  - colorScaleDefaults / colorScaleExplicitThresholds
  - unknownColumnThrows / unknownStyleThrows
  - styleWithNullNameThrows
  - conditionalFormattingNullColumnThrows / NullRuleElementThrows

Then deleted ConditionalFormatsTest.java. Single-file coverage:
33 tests, no overlap.
